set up on Ubuntu 20.04
It was a massive pain to get this stuff working. There were multiple python 2.7 expressions that had to be fixed.

Also, conda is not capable of create new env from the provided yml file. After hours of dependency hell, I managed to create a new yml file with package versions that actually can be installed today on Ubuntu 20.04.

Overall, the pipeline seems working in terms of freesurfer extraction, electrode localization, alignment. But there are bugz for plotting activity on the surface. This needs to be fixed (and will take more time going into the code).